backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
semantic-kernel>=0.4.0
azure-cosmos==4.5.1
pydantic>=2.10.0
pandas>=2.0.0
openpyxl>=3.1.0
python-multipart
pyodbc>=4.0.39
